Update Flightman.go
This script gets data from flightStats API in accordance with the coordinates we sent them. At this stage the coordinates are fixed on the table. Next step would be sending the data from DarkSky API. Last step would be to implement the current date and scope through incoming data for appropriate time frames.